Move block conversion logic to gen-utils, refactor solution state
