+++ src/common/providers/block/reducer/block-reducer.ts
@@ -9,11 +9,9 @@ import {
 } from "../../../../problem-gen/initial-state.ts";
 import { isEqual } from "es-toolkit";
 import {
-  type BlockData,
-  isBlockWithChildrenData,
-  isConstantBlockData,
-} from "../../../block-types.ts";
-import { throwNull } from "../../../utils.ts";
+  convertBlocksToScamper,
+  isConversionError,
+} from "../../../../problem-gen/gen-utils.ts";
 
 export type BlockDispatchType =
   | {
@@ -31,75 +29,6 @@ export type BlockDispatchType =
       type: "CHECK";
     };
 
-interface ConversionError {
-  type: "ConversionError";
-  message: string;
-}
-interface ConversionSuccess {
-  type: "ConversionSuccess";
-  code: string;
-  blocksEncountered: number;
-}
-type ConversionResult = ConversionSuccess | ConversionError;
-
-function isConversionError(
-  result: ConversionResult,
-): result is ConversionError {
-  return result.type === "ConversionError";
-}
-function convertBlocksToScamper(
-  blocks: BlockContextType["blocks"],
-  topLevel: BlockContextType["solutionTopLevel"],
-): ConversionResult {
-  const count = blocks.size;
-
-  function scamperifyBlock(block: BlockData): ConversionResult {
-    if (isConstantBlockData(block)) {
-      return {
-        type: "ConversionSuccess",
-        code: block.value,
-        blocksEncountered: 1,
-      };
-    }
-    if (!isBlockWithChildrenData(block))
-      return throwNull("block is neither constant nor block with children?");
-    // else is block with children
-    const childCode: string[] = [];
-    let blocksEncountered = 1;
-    for (const { id } of block.children) {
-      if (!id) return { type: "ConversionError", message: "null child id" };
-      const conversionResult = scamperifyBlock(
-        blocks.get(id) ?? throwNull("child block id is not a real block?"),
-      );
-      if (isConversionError(conversionResult)) return conversionResult;
-      childCode.push(conversionResult.code);
-      blocksEncountered += conversionResult.blocksEncountered;
-    }
-    return {
-      type: "ConversionSuccess",
-      code: `(${childCode.join(" ")})`,
-      blocksEncountered,
-    };
-  }
-
-  let finalCode: string = "";
-  let blocksEncountered = 0;
-  for (const blockId of topLevel) {
-    const block =
-      blocks.get(blockId) ?? throwNull("top level block not found?");
-    const conversionResult = scamperifyBlock(block);
-    if (isConversionError(conversionResult)) return conversionResult;
-    finalCode += conversionResult.code;
-    blocksEncountered += conversionResult.blocksEncountered;
-  }
-  return blocksEncountered === count
-    ? { type: "ConversionSuccess", code: finalCode, blocksEncountered }
-    : {
-        type: "ConversionError",
-        message: `blocks encountered don't match count (${blocksEncountered.toString()} / ${count.toString()})`,
-      };
-}
-
 export const blockReducer: ImmerReducer<BlockContextType, BlockDispatchType> = (
   draft: Draft<BlockContextType>,
   action: BlockDispatchType,
@@ -115,16 +44,29 @@ export const blockReducer: ImmerReducer<BlockContextType, BlockDispatchType> = (
     }
     case "CHECK": {
       // TODO: make not naive solution checking (probably relies on Scamper integration)
-      console.log(
-        "conversion result: ",
-        convertBlocksToScamper(
-          current(draft).blocks,
-          current(draft).solutionTopLevel,
-        ),
+      const conversionResult = convertBlocksToScamper(
+        current(draft).blocks,
+        current(draft).solution.topLevel,
       );
+      if (isConversionError(conversionResult)) {
+        return {
+          ...draft,
+          solution: {
+            ...current(draft).solution,
+            isCorrect: false,
+          },
+        };
+      }
+      const { code: solutionCode } = conversionResult;
+      console.log(solutionCode);
+
       return {
         ...draft,
-        solutionIsCorrect: isEqual(current(draft).blocks, solutionBlocks),
+        solution: {
+          ...current(draft).solution,
+          isCorrect: isEqual(current(draft).blocks, solutionBlocks),
+          code: solutionCode,
+        },
       };
     }
   }
